+++ src/configdialog.cpp
@@ -10,6 +10,7 @@
 #include "ui_configdialog.h"
 #include "usersdialog.h"
 #include "util.h"
+#include "windowstatestore.h"
 #include <QClipboard>
 #include <QDir>
 #include <QFileDialog>
@@ -37,17 +38,7 @@ ConfigDialog::ConfigDialog(MainWindow *parent)
   mainWindow = parent;
   ui->setupUi(this);
 
-  // Restore dialog state
-  QByteArray savedGeometry = QtPassSettings::getDialogGeometry("configDialog");
-  bool hasSavedGeometry = !savedGeometry.isEmpty();
-  if (hasSavedGeometry) {
-    restoreGeometry(savedGeometry);
-  }
-  if (QtPassSettings::isDialogMaximized("configDialog")) {
-    showMaximized();
-  } else if (!hasSavedGeometry) {
-    // Let window manager handle positioning for first launch
-  }
+  WindowStateStore::restore(*this, QStringLiteral("configDialog"));
 
   const AppSettings s = QtPassSettings::load();
   applySettings(s);
@@ -1158,12 +1149,7 @@ void ConfigDialog::on_checkBoxUseTrayIcon_clicked() {
  * @param event
  */
 void ConfigDialog::closeEvent(QCloseEvent *event) {
-  QtPassSettings::setDialogGeometry("configDialog", saveGeometry());
-  if (!isMaximized()) {
-    QtPassSettings::setDialogPos("configDialog", pos());
-    QtPassSettings::setDialogSize("configDialog", size());
-  }
-  QtPassSettings::setDialogMaximized("configDialog", isMaximized());
+  WindowStateStore::save(*this, QStringLiteral("configDialog"));
   event->accept();
 }