+++ src/qtpasssettings.cpp
@@ -208,35 +208,6 @@ void QtPassSettings::setSavestate(const QByteArray &saveState) {
   getInstance()->setValue(SettingsConstants::savestate, saveState);
 }
 
-auto QtPassSettings::getPos(const QPoint &defaultValue) -> QPoint {
-  QPoint pos =
-      getInstance()->value(SettingsConstants::pos, defaultValue).toPoint();
-  if (pos == QPoint(0, 0)) {
-    QScreen *screen = QGuiApplication::screenAt(QCursor::pos());
-    if (!screen)
-      screen = QGuiApplication::primaryScreen();
-    if (screen)
-      pos = screen->geometry().center();
-  }
-  return pos;
-}
-void QtPassSettings::setPos(const QPoint &pos) {
-  if (pos == QPoint(0, 0))
-    return;
-  getInstance()->setValue(SettingsConstants::pos, pos);
-}
-
-auto QtPassSettings::getSize(const QSize &defaultValue) -> QSize {
-  return getInstance()->value(SettingsConstants::size, defaultValue).toSize();
-}
-void QtPassSettings::setSize(const QSize &size) {
-  getInstance()->setValue(SettingsConstants::size, size);
-}
-
-void QtPassSettings::setMaximized(const bool &maximized) {
-  getInstance()->setValue(SettingsConstants::maximized, maximized);
-}
-
 auto QtPassSettings::getDialogGeometry(const QString &key,
                                        const QByteArray &defaultValue)
     -> QByteArray {
@@ -248,38 +219,6 @@ void QtPassSettings::setDialogGeometry(const QString &key,
                                        const QByteArray &geometry) {
   getInstance()->setValue(SettingsConstants::dialogGeometry + "/" + key,
                           geometry);
-}
-
-auto QtPassSettings::getDialogPos(const QString &key,
-                                  const QPoint &defaultValue) -> QPoint {
-  return getInstance()
-      ->value(SettingsConstants::dialogPos + "/" + key, defaultValue)
-      .toPoint();
-}
-void QtPassSettings::setDialogPos(const QString &key, const QPoint &pos) {
-  getInstance()->setValue(SettingsConstants::dialogPos + "/" + key, pos);
-}
-
-auto QtPassSettings::getDialogSize(const QString &key,
-                                   const QSize &defaultValue) -> QSize {
-  return getInstance()
-      ->value(SettingsConstants::dialogSize + "/" + key, defaultValue)
-      .toSize();
-}
-void QtPassSettings::setDialogSize(const QString &key, const QSize &size) {
-  getInstance()->setValue(SettingsConstants::dialogSize + "/" + key, size);
-}
-
-auto QtPassSettings::isDialogMaximized(const QString &key,
-                                       const bool &defaultValue) -> bool {
-  return getInstance()
-      ->value(SettingsConstants::dialogMaximized + "/" + key, defaultValue)
-      .toBool();
-}
-void QtPassSettings::setDialogMaximized(const QString &key,
-                                        const bool &maximized) {
-  getInstance()->setValue(SettingsConstants::dialogMaximized + "/" + key,
-                          maximized);
 }
 
 void QtPassSettings::setUsePass(const bool &usePass) {