+++ src/usersdialog.cpp
@@ -5,6 +5,7 @@
 #include "pass.h"
 #include "qtpasssettings.h"
 #include "ui_usersdialog.h"
+#include "windowstatestore.h"
 #include <QApplication>
 #include <QCloseEvent>
 #include <QDateTime>
@@ -64,17 +65,7 @@ void UsersDialog::connectSignals() {
  * @brief Restore dialog geometry from settings.
  */
 void UsersDialog::restoreDialogState() {
-  QByteArray savedGeometry = QtPassSettings::getDialogGeometry("usersDialog");
-  bool hasSavedGeometry = !savedGeometry.isEmpty();
-  if (hasSavedGeometry) {
-    restoreGeometry(savedGeometry);
-  }
-  if (QtPassSettings::isDialogMaximized("usersDialog")) {
-    showMaximized();
-  } else if (hasSavedGeometry) {
-    move(QtPassSettings::getDialogPos("usersDialog"));
-    resize(QtPassSettings::getDialogSize("usersDialog"));
-  }
+  WindowStateStore::restore(*this, QStringLiteral("usersDialog"));
 }
 
 auto UsersDialog::loadGpgKeys() -> bool {
@@ -180,12 +171,7 @@ void UsersDialog::accept() {
  * @param event
  */
 void UsersDialog::closeEvent(QCloseEvent *event) {
-  QtPassSettings::setDialogGeometry("usersDialog", saveGeometry());
-  if (!isMaximized()) {
-    QtPassSettings::setDialogPos("usersDialog", pos());
-    QtPassSettings::setDialogSize("usersDialog", size());
-  }
-  QtPassSettings::setDialogMaximized("usersDialog", isMaximized());
+  WindowStateStore::save(*this, QStringLiteral("usersDialog"));
   event->accept();
 }